refactor(gui): 提取窗口居中逻辑至 CenterOnParentMixin，消除5处重复 showEvent
