+++ src/gui/ALConfigWidget.py
@@ -42,6 +42,7 @@ from gui.ALUserTreeWidget import (
     ALUserTreeWidget
 )
 from gui.ALWebDriverDownloadDialog import ALWebDriverDownloadDialog
+from gui.ALWidgetMixin import CenterOnParentMixin
 from gui.resources.ui.Ui_ALConfigWidget import Ui_ALConfigWidget
 from interfaces.ConfigProvider import (
     CfgKey,
@@ -52,7 +53,7 @@ from utils.JSONReader import JSONReader
 from utils.JSONWriter import JSONWriter
 
 
-class ALConfigWidget(QWidget, Ui_ALConfigWidget):
+class ALConfigWidget(CenterOnParentMixin, QWidget, Ui_ALConfigWidget):
 
     configWidgetIsClosed = Signal()
 
@@ -109,29 +110,6 @@ class ALConfigWidget(QWidget, Ui_ALConfigWidget):
         self.LoadConfigButton.clicked.connect(self.onLoadConfigButtonClicked)
         self.ConfirmButton.clicked.connect(self.onConfirmButtonClicked)
         self.CancelButton.clicked.connect(self.onCancelButtonClicked)
-
-    def showEvent(
-        self,
-        event
-    ):
-
-        result = super().showEvent(event)
-
-        screen_rect = self.screen().geometry()
-        target_pos = self.parent().geometry().center()
-        target_pos.setX(target_pos.x() - self.width()//2)
-        target_pos.setY(target_pos.y() - self.height()//2)
-        if target_pos.x() < 0:
-            target_pos.setX(0)
-        if target_pos.x() + self.width() > screen_rect.width():
-            target_pos.setX(screen_rect.width() - self.width())
-        if target_pos.y() < 0:
-            target_pos.setY(0)
-        if target_pos.y() + self.height() > screen_rect.height():
-            target_pos.setY(screen_rect.height() - self.height())
-        self.move(target_pos)
-
-        return result
 
     def closeEvent(
         self,
