+++ src/gui/ALWebDriverDownloadDialog.py
@@ -38,6 +38,7 @@ from managers.driver.WebDriverManager import (
     WebDriverStatus
 )
 from gui.ALStatusLabel import ALStatusLabel
+from gui.ALWidgetMixin import CenterOnParentMixin
 
 
 class DownloadWorker(QThread):
@@ -123,7 +124,7 @@ class DownloadWorker(QThread):
                 self.wait()
 
 
-class ALWebDriverDownloadDialog(QDialog):
+class ALWebDriverDownloadDialog(CenterOnParentMixin, QDialog):
 
     def __init__(
         self,
@@ -151,28 +152,6 @@ class ALWebDriverDownloadDialog(QDialog):
         self.connectSignals()
         self.initializeDriverManager()
         self.refreshDriverList()
-
-    def showEvent(
-        self,
-        event
-    ):
-
-        result = super().showEvent(event)
-        if self.parent():
-            screen_rect = self.screen().geometry()
-            target_pos = self.parent().geometry().center()
-            target_pos.setX(target_pos.x() - self.width()//2)
-            target_pos.setY(target_pos.y() - self.height()//2)
-            if target_pos.x() < 0:
-                target_pos.setX(0)
-            if target_pos.x() + self.width() > screen_rect.width():
-                target_pos.setX(screen_rect.width() - self.width())
-            if target_pos.y() < 0:
-                target_pos.setY(0)
-            if target_pos.y() + self.height() > screen_rect.height():
-                target_pos.setY(screen_rect.height() - self.height())
-            self.move(target_pos)
-        return result
 
     def setupUi(
         self
